Fixed missing metadata, added auto-update search
On some sites like southpark, metadata like the video duration is in an entry and not in the main JSON information. When information is empty, OnionMedia looks into the entries if metadata is available there.
The Microsoft Store version of OnionMedia checks now for updates on app start (Package.appinstaller)